Hey Eric,
I just had a few minutes to finish up the function, it runs ok, so I am not going to spend too much more time on it tonight as I have to head out...

Let me know if it works for you, and give me all your suggestions to improve it.

Thanks for showing me all that stuff on R, I am really glad to learn so much!

Have a good evening,

François